redis 대기열 생성 - 초안 -TODO : 1)스케줄러분산락 2)주석 삭제 3)token dto 입력

+++ src/main/java/io/hhplus/concertbook/domain/dto/TokenDto.java
@@ -9,6 +9,6 @@ public class TokenDto {
     String token;
     String userLoginId;
     ApiNo apiNo;
-    int waitNo;
+    long waitNo;
     WaitStatus waitStatus;
 }
